test_ver

+++ src/openne/models/grarep.py
@@ -31,7 +31,7 @@ class GraRep(ModelWithEmbeddings):
         return torch.as_tensor(Ud)*torch.pow(Sd, alpha)
 
     def get_train(self, graph, **kwargs):
-        adj = torch.from_numpy(graph.adjmat(directed=False, weighted=False, scaled=1))
+        adj = torch.from_numpy(graph.adjmat(directed=False, weighted=False, scaled=1)).type(torch.float32)
         Ak = torch.eye(graph.nodesize)
         RepMat = torch.zeros((graph.nodesize, int(self.dim * self.kstep)))
         for i in range(self.kstep):
